RegulationFixes, Car improvements
Allowed to evolve (only module active) the children of regulation
modules (no changes were required actually).

Fixed a bug with regulation modules (and modules in general) where IDs
were not reserved for new local outputs (crucial in regulation modules,
affected with 3 children and more).

Added new tracks and track selection in cars.

Added "cancel" option when adding modules.

Fixed an issue where champions were not destroyed before evolution
(because of a delay in a IEnumerator).

Time now is set to the slider values when starting and stopping
evolution.